As outlined in #80, Koanf keeps track of KeyMap internally even though the representation is not really needed. KeyMaps use a significant amount of memory depending on the config layout which can lead to excessive memory consumptions for large or complex configs.

This patch removes the need to keep keyMap around and therefore significantly improves performance:

```
benchmark                old ns/op     new ns/op     delta
BenchmarkLoadFile-16     146821        135027        -8.03%

benchmark                old allocs     new allocs     delta
BenchmarkLoadFile-16     679            657            -3.24%

benchmark                old bytes     new bytes     delta
BenchmarkLoadFile-16     46277         37019         -20.01%
```

BREAKING CHANGE: This PR deprecates `KeyMap` function. It still works as before but is slower.

It also removes the second return argument of `maps.Flatten` which was the KeyMap previously.