@knadh 2021-06-08
This is a significant change, but we can consider it. The idea of keymap was to sacrifice a bit of performance on load, but keep all subsequent access fast. Thus, throwing away key map results in a recursive search `maps.Search()` of every item. The more nested it is, the worse the performance.

Here's a comparison of `master` with this PR. The benchmark is:

```go
func BenchmarkGet(b *testing.B) {
	k := koanf.New(".")
	k.Load(file.Provider(mockYAML), yaml.Parser())

	b.ReportAllocs()
	for n := 0; n < b.N; n++ {
		k.Bool("parent1.child1.grandchild1.on")
	}
}
```

### master

```go
$ go test -bench=.                                                                                                                                                                        
goos: linux
goarch: amd64
pkg: github.com/knadh/koanf
cpu: Intel(R) Core(TM) i7-10510U CPU @ 1.80GHz
BenchmarkGet-8          17975173                65.18 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/knadh/koanf  6.159s
```

### This PR

```go
$ go test -bench=.                                                                                                                                                                        
goos: linux
goarch: amd64
pkg: github.com/knadh/koanf
cpu: Intel(R) Core(TM) i7-10510U CPU @ 1.80GHz
BenchmarkGet-8           5147134               197.4 ns/op            80 B/op          2 allocs/op
PASS
ok      github.com/knadh/koanf  5.312s
```