@aeneasr 2021-06-08
Ah I see - I assume this performance coming from `strings.Split()` probably? I guess repeated access outweighs an expensive write here.

I think the overall memory problem is probably the recursion because the GC keeps the parent's functions variables around (these are often copied). If you look at the execution tree of the recursive `map.flatten` in the flamegraph you see that the lower most invokation has the most memory allocated because it inherits all memory from all parents.

If instead `maps.Flatten` was iterative this memory allocation could be much improved, allowing the GC to free up memory as we no longer need it.

Alternatively a channel could be used to still traverse the tree recursively but send the results on the channel and then free the variables to free up memory.