@knadh 2021-06-08
Yeah. We should optimise `maps.Flatten()` first and then figure out if KeyMap is worth keeping. I'll give this a shot. In the meanwhile if you have anything, please do share.