@knadh 2021-06-09
I managed to write an (inefficient) non-recursive prototype last night. https://gist.github.com/knadh/9520b2a3f8edf589c450ed7e283ba60f

```go
$ go test -bench=. *.go                                                                                                                                                                   
goos: linux
goarch: amd64
cpu: Intel(R) Core(TM) i7-10510U CPU @ 1.80GHz
BenchmarkFlattenRecurse-8         302734              3332 ns/op            2815 B/op         31 allocs/op
BenchmarkFlattenIterate-8         353250              3544 ns/op            2784 B/op         28 allocs/op
PASS
ok      command-line-arguments  2.339s
```