@aeneasr 2021-06-09
Nice! :) You can also observe the raw performance by doing:

```
		b.StopTimer()
		m := CopyMap(nested)
		b.StartTimer()
		Flatten...
```

which removes the CopyMap allocs / exec time from the benchmark results!