@knadh 2021-06-09
Ah yes, forgot about that. Updated the gist.

```shell
goos: linux
goarch: amd64
cpu: Intel(R) Core(TM) i7-10510U CPU @ 1.80GHz
BenchmarkFlattenRecurse-8         482095              2252 ns/op             464 B/op         17 allocs/op
BenchmarkFlattenIterate-8         504212              2551 ns/op             432 B/op         14 allocs/op
PASS
ok      command-line-arguments  14.341s
```