@rhnvrm 2021-06-09
Hey

I tried refactoring/rewriting the logic for FlattenRecurse. You can find the gist here: https://gist.github.com/rhnvrm/006d5095665f54eadd0f8c5f238d82cb

Here are the benchmarks:

```
❯ go test -bench=. -benchmem
goos: linux
goarch: amd64
pkg: github.com/knadh/koanf/gg
cpu: Intel(R) Core(TM) i7-8665U CPU @ 1.90GHz
BenchmarkFlattenRecurseOld-8   	  494610	      2459 ns/op	     464 B/op	      17 allocs/op
BenchmarkFlattenRecurse-8      	  880119	      1420 ns/op	      48 B/op	       9 allocs/op
BenchmarkFlattenIterate-8      	  431438	      2863 ns/op	     432 B/op	      14 allocs/op
PASS
ok  	github.com/knadh/koanf/gg	29.910s
```