@rhnvrm 2021-06-14
Hey

I tried adding the keyMap creation to my refactored recursive approach above but creating the slices across the stack produces lots of allocs. 

I gave more thought to this over the weekend.

We have two approaches at the moment, both with their unique trade-offs.

1. Optimize for loading
2. Optimize for lookups

Currently, koanf is more optimally doing lookups but does a trade-off for loading with increased memory usage. At the same time, if we switch out the current approach for better load performance, we might see lower perf for usecases with large amounts of lookups. 

Maybe we can consider, a different alternative along the lines of lazy loading the map or caching the string.Split() results. 

1. We create and start building the keyMap as a cache instead of on boot; on demand. This might require to hook onChange to delete specific keys.
2. A simpler way could be to simply cache string.Split() results and the paths. The benefit here is that we can simply delete the map onChange.

These approaches might return higher lookup times for the first lookup but be faster for significant further lookups. At the same time it can avoid loading and maintaining the entire keyMap for all the koanf paths. Thereby being a compromise for both the desired outcomes for the library.

@aeneasr, @knadh, Please feel free to give your inputs and feedback on this approach and what you think might be the impact of this based on your recent benchmarks and production usecases.